Refactored deliverables to cards
