+++ src/components/DeliverableView/QuestionsDeliverableCard.tsx
@@ -1,0 +1,104 @@
+import React, { useEffect, useRef } from 'react';
+
+import { Box, Typography, useTheme } from '@mui/material';
+import { Message } from '@terraware/web-components';
+
+import Metadata from './Metadata';
+import { EditProps } from './types';
+import DeliverableDisplayVariableValue from 'src/components/DocumentProducer/DeliverableDisplayVariableValue';
+import Card from 'src/components/common/Card';
+import { useIsVisible } from 'src/hooks/useIsVisible';
+import { requestListDeliverableVariablesValues } from 'src/redux/features/documentProducer/values/valuesThunks';
+import { selectDeliverableVariablesWithValues } from 'src/redux/features/documentProducer/variables/variablesSelector';
+import { requestListDeliverableVariables } from 'src/redux/features/documentProducer/variables/variablesThunks';
+import { useAppDispatch, useAppSelector } from 'src/redux/store';
+import { VariableWithValues } from 'src/types/documentProducer/Variable';
+import { VariableValue } from 'src/types/documentProducer/VariableValue';
+import { variableDependencyMet } from 'src/utils/documentProducer/variables';
+
+import QuestionsDeliverableStatusMessage from './QuestionsDeliverableStatusMessage';
+
+const QuestionBox = ({ projectId, variable }: { projectId: number; variable: VariableWithValues }): JSX.Element => {
+  const theme = useTheme();
+  const containerRef = useRef(null);
+  const visible = useIsVisible(containerRef);
+
+  const firstVariableValue: VariableValue | undefined = (variable?.variableValues || [])[0];
+
+  return (
+    <Box
+      className={visible ? 'question-visible' : undefined}
+      data-variable-id={variable.id}
+      ref={containerRef}
+      sx={{ marginBottom: theme.spacing(4) }}
+    >
+      <Box sx={{ float: 'right', marginBottom: '16px', marginLeft: '16px' }}>
+        {/* <DeliverableStatusBadge status={variableWithValues.status} /> */}
+      </Box>
+      <Typography sx={{ fontWeight: '600', marginBottom: '16px' }}>{variable.name}</Typography>
+      {!!variable.description && (
+        <Typography
+          sx={{
+            color: theme.palette.TwClrTxtSecondary,
+            fontSize: '14px',
+            fontStyle: 'italic',
+            fontWeight: 400,
+            lineHeight: '20px',
+            marginBottom: '16px',
+          }}
+        >
+          {variable.description}
+        </Typography>
+      )}
+      {!!firstVariableValue?.feedback && (
+        <Box marginBottom={theme.spacing(2)}>
+          <Message body={firstVariableValue.feedback} priority='critical' type='page' />
+        </Box>
+      )}
+      <DeliverableDisplayVariableValue projectId={projectId} variable={variable} />
+    </Box>
+  );
+};
+
+const QuestionsDeliverableCard = ({ deliverable }: EditProps): JSX.Element | null => {
+  const dispatch = useAppDispatch();
+  const theme = useTheme();
+
+  const variablesWithValues: VariableWithValues[] = useAppSelector((state) =>
+    selectDeliverableVariablesWithValues(state, deliverable.id, deliverable.projectId)
+  );
+
+
+  useEffect(() => {
+    void dispatch(requestListDeliverableVariables(deliverable.id));
+    void dispatch(requestListDeliverableVariablesValues({ deliverableId: deliverable.id, projectId: deliverable.projectId }));
+  }, [deliverable]);
+
+  if (!deliverable) {
+    return null;
+  }
+
+  return (
+    <Box display='flex' flexDirection='column' flexGrow={1}>
+      <QuestionsDeliverableStatusMessage deliverable={deliverable} variables={variablesWithValues} />
+      <Card style={{ display: 'flex', flexDirection: 'column', flexGrow: 1 }}>
+        <Metadata deliverable={deliverable} />
+        <Box
+          sx={{
+            borderTop: `1px solid ${theme.palette.TwClrBrdrTertiary}`,
+            marginBottom: theme.spacing(4),
+            paddingTop: theme.spacing(3),
+          }}
+        >
+          {variablesWithValues.map((variableWithValues: VariableWithValues, index: number) =>
+            variableDependencyMet(variableWithValues, variablesWithValues) ? (
+              <QuestionBox key={index} projectId={deliverable.projectId} variable={variableWithValues} />
+            ) : null
+          )}
+        </Box>
+      </Card>
+    </Box>
+  );
+};
+
+export default QuestionsDeliverableCard;
