+++ src/components/DeliverableView/types.ts
@@ -4,6 +4,7 @@ import { Deliverable } from 'src/types/Deliverables';
 
 export type ViewProps = {
   deliverable: Deliverable;
+  setSubmitButtonDisalbed?: (disabled: boolean) => void;
 };
 
 export type EditProps = ViewProps & {
