+++ src/scenes/DeliverablesRouter/DeliverablePage.tsx
@@ -1,0 +1,30 @@
+import React, { ReactNode } from "react";
+
+import { Crumb } from "src/components/BreadCrumbs";
+import TitleBar from "src/components/DeliverableView/TitleBar"
+import { EditProps } from "src/components/DeliverableView/types"
+import Page from "src/components/Page"
+import { useDeviceInfo } from "@terraware/web-components/utils";
+import MobileMessage from "src/components/DeliverableView/MobileMessage";
+
+type DeliverablePageProp = EditProps & {
+  children?: ReactNode;
+  crumbs?: Crumb[];
+  isLoading?: boolean;
+  rightComponent?: ReactNode;
+} 
+
+const DeliverablePage = (props: DeliverablePageProp) => {
+  const { children, crumbs, isLoading, rightComponent } = props;
+  const { isMobile } = useDeviceInfo();
+
+  if (isMobile) {
+    return <MobileMessage {...props} />;
+  }
+
+  return <Page title={<TitleBar {...props} />} crumbs={crumbs} isLoading={isLoading} rightComponent={rightComponent}>
+    {children}
+  </Page>
+}
+
+export default DeliverablePage;
