+++ src/scenes/DeliverablesRouter/index.tsx
@@ -5,9 +5,9 @@ import DeliverableProvider from 'src/providers/Deliverable/DeliverableProvider';
 import ParticipantProvider from 'src/providers/Participant/ParticipantProvider';
 import ProjectProvider from 'src/providers/Project/ProjectProvider';
 
-import DeliverableViewWrapper from './DeliverableViewWrapper';
 import DeliverablesList from './DeliverablesList';
 import QuestionsDeliverableEdit from './QuestionsDeliverableEditView';
+import DeliverableView from './DeliverableView';
 
 const DeliverablesRouter = (): JSX.Element => {
   return (
@@ -18,7 +18,7 @@ const DeliverablesRouter = (): JSX.Element => {
           <ProjectProvider>
             <ParticipantProvider>
               <DeliverableProvider>
-                <DeliverableViewWrapper />
+                <DeliverableView />
               </DeliverableProvider>
             </ParticipantProvider>
           </ProjectProvider>
